(PC-20298)[PRO] refactor: use router v6 and rename url for signupJourney
