+++ pro/src/app/AppRouter/routes_map.tsx
@@ -26,7 +26,7 @@ import SetPassword from 'pages/SetPassword/SetPassword'
 import SetPasswordConfirm from 'pages/SetPasswordConfirm/SetPasswordConfirm'
 import SignIn from 'pages/SignIn/SignIn'
 import Signup from 'pages/Signup/Signup'
-import { SignupJourneyRoutes } from 'pages/SignupOffererRoutes'
+import { SignupJourneyRoutes } from 'pages/SignupJourneyRoutes'
 import { UserProfile } from 'pages/User'
 import { VenueCreation } from 'pages/VenueCreation'
 import { VenueEdition } from 'pages/VenueEdition'
@@ -339,13 +339,9 @@ const routes: IRoute[] = [
   },
   {
     component: SignupJourneyRoutes,
-    path: [
-      '/signup/authentification',
-      '/signup/activite',
-      '/signup/validation',
-    ],
-    exact: true,
+    path: '/parcours-inscription/*',
     title: 'Parcours de souscription',
+    useV6Router: true,
   },
 ]
